feat: OCR Image Scan
Implemented the feature for extracting all the text from an attached image
This feature is only kept as additional or bonus feature, as the main feature is text extraction by field

Also, this includes a combobox in the scan form for language selection. The backend expects a language
definition in the request defining the language the text in the document is written in.